@@ -20,7 +20,7 @@ classifiers =
 
 [options]
 zip_safe = False
-python_requires = >= 3.8
+python_requires = >= 3.9
 install_requires =
     numpy
     networkx
