@@ -2,6 +2,8 @@
 # For details, see https://github.com/OpenFreeEnergy/openfe
 
 import pathlib
+import logging
+import logging.config
 
 import click
 from plugcli.cli import CLI, CONTEXT_SETTINGS
@@ -9,15 +11,19 @@ from plugcli.plugin_management import FilePluginLoader
 
 import openfecli
 
-from .plugins import OFECommandPlugin
+from openfecli.plugins import OFECommandPlugin
 
 
 class OpenFECLI(CLI):
     COMMAND_SECTIONS = ["Setup", "Simulation", "Orchestration", "Analysis"]
 
-    def get_installed_plugins(self):
+    def get_loaders(self):
         commands = str(pathlib.Path(__file__).parent.resolve() / "commands")
         loader = FilePluginLoader(commands, OFECommandPlugin)
+        return [loader]
+
+    def get_installed_plugins(self):
+        loader = self.get_loaders()[0]
         return list(loader())
 
 
@@ -30,11 +36,13 @@ the OpenFE Python library.
 @click.command(cls=OpenFECLI, name="openfe", help=_MAIN_HELP,
                context_settings=CONTEXT_SETTINGS)
 @click.version_option(version=openfecli.__version__)
-def main():
-    # currently empty: we can add options at the openfe level (as opposed to
-    # subcommands) by adding click options here. Subcommand runs after this
-    # is the processed.
-    pass
+@click.option('--log', type=click.Path(exists=True, readable=True),
+              help="logging configuration file")
+def main(log):
+    # Subcommand runs after this is processed.
+    # set logging if provided
+    if log:
+        logging.config.fileConfig(log, disable_existing_loggers=False)
 
 
 if __name__ == "__main__":  # -no-cov- (useful in debugging)
